@@ -1,13 +1,14 @@
 {
   "name": "redux-test-template",
-  "version": "0.0.2",
+  "version": "0.0.3",
   "description": "Redux Express Webpack Isomorphic ",
   "main": "server.js",
   "scripts": {
-    "start:x": "npm run build && export UNIVERSAL=true && babel-node src/run-server.js",
-    "start": "npm run build && babel-node src/run-server.js",
-    "start:dev": "export NODE_ENV=development && npm run build:dev && nodemon --exec babel-node -- src/run-server.js",
+    "start": "npm run build && node server/run-server.js",
+    "start:aws": "export NODE_ENV=production && npm rebuild node-sass && webpack && node server/run-server.js",
+    "start:dev": "export NODE_ENV=development && npm run build:dev && nodemon --exec node -- server/run-server.js",
     "start:universal": "export UNIVERSAL=true && npm run start",
+    "start:universal:aws": "export NODE_ENV=production && export UNIVERSAL=true && npm run start:aws",
     "start:dev:universal": "export NODE_ENV=development && export UNIVERSAL=true && npm run start:dev",
     "build": "NODE_ENV=production webpack -p",
     "build:dev": "webpack -d",
@@ -17,59 +18,59 @@
   "author": "JT Riddick",
   "license": "ISC",
   "dependencies": {
-    "axios": "^0.16.2",
     "babel-preset-stage-0": "^6.24.1",
     "babel-preset-stage-1": "^6.24.1",
     "ejs": "^2.5.6",
     "express": "^4.15.3",
     "lodash": "^4.17.4",
     "prop-types": "^15.5.10",
-    "react": "^15.5.4",
-    "react-dom": "^15.5.4",
+    "react": "^15.6.1",
+    "react-bootstrap": "^0.31.0",
+    "react-dom": "^15.6.1",
     "react-redux": "^5.0.5",
+    "react-router-bootstrap": "^0.24.2",
     "react-router-dom": "^4.1.1",
-    "redux": "^3.6.0",
-    "redux-form": "^6.8.0",
-    "redux-promise": "^0.5.3"
+    "redux": "^3.7.0",
+    "redux-devtools-extension": "^2.13.2",
+    "redux-promise": "^0.5.3",
+    "redux-thunk": "^2.2.0"
   },
   "devDependencies": {
     "babel-cli": "^6.24.1",
     "babel-core": "^6.25.0",
-    "babel-loader": "^7.0.0",
+    "babel-loader": "^7.1.0",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-preset-env": "^1.5.2",
     "babel-preset-es2015": "^6.24.1",
-    "babel-preset-es2015-node5": "^1.2.0",
+    "babel-preset-es2015-node6": "^0.4.0",
     "babel-preset-react": "^6.24.1",
-    "babel-preset-react-app": "^3.0.0",
     "chai": "^4.0.2",
     "chalk": "^1.1.3",
     "css-loader": "^0.28.4",
-    "eslint": "^4.0.0",
-    "eslint-config-airbnb": "^15.0.1",
-    "eslint-plugin-import": "^2.3.0",
-    "eslint-plugin-jsx-a11y": "^5.0.3",
-    "eslint-plugin-react": "^7.1.0",
     "extract-text-webpack-plugin": "^2.1.2",
+    "file-loader": "^0.11.2",
     "http-server": "^0.10.0",
     "jsdom": "^11.0.0",
     "mocha": "^3.4.2",
     "morgan": "^1.8.2",
     "node-sass": "^4.5.3",
     "nodemon": "^1.11.0",
-    "postcss-loader": "^2.0.5",
-    "promise": "^7.1.1",
-    "react-addons-test-utils": "^15.5.1",
-    "sass-loader": "^6.0.5",
+    "postcss-loader": "^2.0.6",
+    "promise": "^8.0.0",
+    "react-addons-test-utils": "^15.6.0",
+    "sass-loader": "^6.0.6",
     "style-loader": "^0.18.2",
     "uglify-js": "git://github.com/mishoo/UglifyJS2#harmony-v2.8.22",
-    "uglifyjs-webpack-plugin": "^0.4.3",
-    "webpack": "^2.6.1",
+    "uglifyjs-webpack-plugin": "^0.4.6",
+    "url-loader": "^0.5.9",
+    "webpack": "^3.0.0",
     "webpack-split-by-path": "^2.0.0"
   },
   "babel": {
     "presets": [
-      "react-app"
+      "react",
+      "es2015",
+      "stage-2"
     ]
   },
   "eslintConfig": {
