@@ -1,9 +1,10 @@
 // Sourcemap babel-loader scss loader uglifyjs hmr hml
 import webpack from 'webpack';
 import path from 'path';
-const UglifyJSPlugin = require('uglifyjs-webpack-plugin');
-const ExtractTextPlugin = require('extract-text-webpack-plugin');
-const SplitByPathPlugin = require('webpack-split-by-path');
+import UglifyJSPlugin from 'uglifyjs-webpack-plugin';
+import ExtractTextPlugin from 'extract-text-webpack-plugin';
+import SplitByPathPlugin from 'webpack-split-by-path';
+import autoprefixer from 'autoprefixer';
 
 require.extensions['.scss'] = () => { return; }; require.extensions['.css'] = () => { return; };
 
@@ -22,7 +23,7 @@ const config = {
     js: './src/app-client.js',
   },
   output: {
-    path: path.join(__dirname, 'src', 'static', 'js'),
+    path: path.join(__dirname, 'src', 'public', 'js'),
     filename: 'bundle.js',
   },
   module: {
@@ -40,18 +41,35 @@ const config = {
         test:/(\.scss$|.css$)/,
         include: path.join(__dirname, 'src','sass'),
         exclude: /node_modules/,
-        loader: extractCSS.extract('css-loader?modules=true!sass-loader?sourceMap=true?')
+        //postcss loader goes before sass-loader
+        loader: extractCSS.extract('css-loader?modules&importLoaders=1&localIdentName=[name]__[local]___[hash:base64:5]!sass-loader?sourceMap=true')
+      },
+      {
+        test:/(\.png$|.jpg$|.gif$)/,
+        use:[{
+          loader: 'url-loader',
+          options: { limit:10000 }
+        }]
       },
     ],
   },
   plugins:[
-    new webpack.SourceMapDevToolPlugin({
-      // 404 from morgan with this filename
-      // filename: 'bundle.js.map',
-      exclude: /node_modules/,
+    new webpack.LoaderOptionsPlugin({
+      options: {
+        postcss: [
+          autoprefixer({
+            browsers: [
+              'last 3 version',
+              'ie >= 10',
+            ],
+          }),
+        ],
+        context: path.join(__dirname, 'src'),
+      },
     }),
     extractCSS,
     extractSCSS,
+    new webpack.HotModuleReplacementPlugin(),
     new webpack.DefinePlugin({
       'process.env.NODE_ENV' : JSON.stringify(process.env.NODE_ENV)
     }),
