@@ -1,17 +1,15 @@
 import React, { Component } from 'react';
+import Clicker from './Clicker.js';
 
 import style from '../sass/style.scss';
 
 
 export default class Index extends Component {
-
-  componentDidMount(){
-    console.log('index test log : ', this);
-  }
-
   render() {
     return (
-      <div className={style.test}>React simple starter</div>
+      <div className={style.test}>
+        <h1>I am a heading</h1>
+      </div>
     );
   }
 }
