@@ -1,7 +1,12 @@
 import { combineReducers } from 'redux';
+import clickReducer from './click_reducer';
+import cardReducer from './card_reducer';
+
 
 const rootReducer = combineReducers({
-  state: (state = {}) => state
+  // state: (state = {}) => state
+  number: clickReducer,
+  card: cardReducer,
 });
 
 export default rootReducer;
